backend/python-service/requirements.txt
# Browser-Use AI Agent
browser-use>=0.1.40
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0

# Browser Automation
playwright>=1.48.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
google-generativeai>=0.8.0
